arregla bugs de formularios ABM

+++ GUILayer/Transacciones/Frm_buscar_Atencion.vb
@@ -17,8 +17,6 @@ Public Class frm_buscar_atencion
         Me.habilitarControles()
 
         'Estado inicial de los filtros
-        'dtp_fechaInicio.Value = Convert.ToDateTime("02/02/2002")
-        'dtp_fechaFinal.Value = Convert.ToDateTime("02/02/2002")
         dtp_fechaInicio.Enabled = False
         dtp_fechaFinal.Enabled = False
         chk_usoFiltros.CheckState = CheckState.Unchecked
@@ -68,10 +66,15 @@ Public Class frm_buscar_atencion
     End Sub
 
     Private Sub btn_buscar2_Click(sender As Object, e As EventArgs) Handles btn_buscar2.Click
-        'Faltan validar las fechas 
+        'frm_registrar_Atencion.busqueda = False (El formulario es llamado desde Frm_registrar_Atencion)
         'Si se cuenta con el tipo y nro de documento llena la grilla 
         If frm_registrar_Atencion.busqueda = False Then
             If usoFiltrosFecha = True Then
+                If (dtp_fechaInicio.Value > dtp_fechaFinal.Value) Then
+                    MessageBox.Show("Periodo de fechas incorrecto", "Aviso", MessageBoxButtons.OK, MessageBoxIcon.Error)
+                    Exit Sub
+                End If
+
                 If (cbo_tiposDocumento.SelectedIndex <> -1 And txt_docAfiliado.Text <> String.Empty) = True Then
                     Me.llenarGrilla(Nothing, dtp_fechaInicio.Value.ToString("dd/MM/yyyy"), dtp_fechaFinal.Value.ToString("dd/MM/yyyy"), Integer.Parse(cbo_tiposDocumento.SelectedValue), Convert.ToInt32(txt_docAfiliado.Text))
                 Else
@@ -90,9 +93,14 @@ Public Class frm_buscar_atencion
 
             End If
         Else
+            'frm_registrar_Atencion.busqueda = True (El formulario es llamado desde el Menú)
             'Si no hay datos se llena la grilla con todas las atenciones 
             If frm_registrar_Atencion.busqueda = True And (cbo_tiposDocumento.SelectedIndex <> -1 And txt_docAfiliado.Text <> String.Empty) = True Then
                 If usoFiltrosFecha = True Then
+                    If (dtp_fechaInicio.Value > dtp_fechaFinal.Value) Then
+                        MessageBox.Show("Periodo de fechas incorrecto", "Aviso", MessageBoxButtons.OK, MessageBoxIcon.Error)
+                        Exit Sub
+                    End If
                     Me.llenarGrilla(Nothing, dtp_fechaInicio.Value, dtp_fechaFinal.Value, Integer.Parse(cbo_tiposDocumento.SelectedValue), Convert.ToInt32(txt_docAfiliado.Text))
                 Else
                     Me.llenarGrilla(Nothing, Nothing, Nothing, Integer.Parse(cbo_tiposDocumento.SelectedValue), Convert.ToInt32(txt_docAfiliado.Text))
@@ -208,23 +216,25 @@ Public Class frm_buscar_atencion
 
     Private Sub btn_eliminar_Click(sender As Object, e As EventArgs) Handles btn_eliminar.Click
         If Convert.ToInt16(row_selected.Cells(14).Value.ToString()) <> 3 Then
-            Dim oAtencion As New Atencion
-            With oAtencion
-                .fechaAtencion = Convert.ToDateTime(row_selected.Cells(0).Value.ToString())
-                .tipoDocAfiliado = Convert.ToInt16(row_selected.Cells(2).Value.ToString())
-                .nroDocAfiliado = Convert.ToInt64(row_selected.Cells(3).Value.ToString())
-                .idCentro = Convert.ToInt16(row_selected.Cells(7).Value.ToString())
-                .matriculaProfesional = Convert.ToInt64(row_selected.Cells(9).Value.ToString())
-                .idEspecialidad = Convert.ToInt16(row_selected.Cells(11).Value.ToString())
-                .fechaAltaProfesional = Convert.ToDateTime(row_selected.Cells(13).Value.ToString())
-            End With
-            If oAtencionService.eliminarAtencion(oAtencion) = True Then
-                MessageBox.Show("La Atención se eliminó con éxito", "Aviso", MessageBoxButtons.OKCancel, MessageBoxIcon.Exclamation _
-                               , MessageBoxDefaultButton.Button1)
-                Me.llenarGrilla(Nothing, Nothing, Nothing, Integer.Parse(cbo_tiposDocumento.SelectedValue), Convert.ToInt32(txt_docAfiliado.Text))
-            Else
-                MessageBox.Show("Hubo un problema en la eliminación de la Atención", "Aviso", MessageBoxButtons.OKCancel, MessageBoxIcon.Exclamation _
-                              , MessageBoxDefaultButton.Button1)
+            If MessageBox.Show("Esta seguro que desea eliminar la Atención seleccionada?", "Aviso", MessageBoxButtons.OKCancel, MessageBoxIcon.Question, MessageBoxDefaultButton.Button1) = Windows.Forms.DialogResult.OK Then
+                Dim oAtencion As New Atencion
+                With oAtencion
+                    .fechaAtencion = Convert.ToDateTime(row_selected.Cells(0).Value.ToString())
+                    .tipoDocAfiliado = Convert.ToInt16(row_selected.Cells(2).Value.ToString())
+                    .nroDocAfiliado = Convert.ToInt64(row_selected.Cells(3).Value.ToString())
+                    .idCentro = Convert.ToInt16(row_selected.Cells(7).Value.ToString())
+                    .matriculaProfesional = Convert.ToInt64(row_selected.Cells(9).Value.ToString())
+                    .idEspecialidad = Convert.ToInt16(row_selected.Cells(11).Value.ToString())
+                    .fechaAltaProfesional = Convert.ToDateTime(row_selected.Cells(13).Value.ToString())
+                End With
+                If oAtencionService.eliminarAtencion(oAtencion) = True Then
+                    MessageBox.Show("La Atención se eliminó con éxito", "Aviso", MessageBoxButtons.OKCancel, MessageBoxIcon.Exclamation _
+                                   , MessageBoxDefaultButton.Button1)
+                    Me.llenarGrilla(Nothing, Nothing, Nothing, Integer.Parse(cbo_tiposDocumento.SelectedValue), Convert.ToInt32(txt_docAfiliado.Text))
+                Else
+                    MessageBox.Show("Hubo un problema en la eliminación de la Atención", "Aviso", MessageBoxButtons.OKCancel, MessageBoxIcon.Exclamation _
+                                  , MessageBoxDefaultButton.Button1)
+                End If
             End If
         Else
             MessageBox.Show("La Atención seleccionada ya se encuentra anulada", "Aviso", MessageBoxButtons.OKCancel, MessageBoxIcon.Error _
